+++ packages/gamut-styles/src/variance/config.ts
@@ -89,6 +89,13 @@ export const border = {
   borderStyleBottom: { property: 'borderBottomStyle' },
 } as const;
 
+const alignments = {
+  justifyContent: { property: 'justifyContent' },
+  justifyItems: { property: 'justifyItems' },
+  alignItems: { property: 'alignItems' },
+  alignContent: { property: 'alignContent' },
+} as const;
+
 export const flex = {
   flex: { property: 'flex' },
   flexBasis: { property: 'flexBasis' },
@@ -97,10 +104,7 @@ export const flex = {
   flexShrink: { property: 'flexShrink' },
   flexGrow: { property: 'flexGrow' },
   order: { property: 'order' },
-  justifyContent: { property: 'justifyContent' },
-  justifyItems: { property: 'justifyItems' },
-  alignItems: { property: 'alignItems' },
-  alignContent: { property: 'alignContent' },
+  ...alignments,
 } as const;
 
 export const grid = {
@@ -119,6 +123,7 @@ export const grid = {
   gridRowStart: { property: 'gridRowStart' },
   gridColumnEnd: { property: 'gridColumnEnd' },
   gridRowEnd: { property: 'gridRowEnd' },
+  ...alignments,
 } as const;
 
 export const background = {
@@ -149,6 +154,11 @@ export const layout = {
   overflow: { property: 'overflow' },
   overflowX: { property: 'overflowX' },
   overflowY: { property: 'overflowY' },
+  size: {
+    property: 'width',
+    properties: ['width', 'height'],
+    transform: transformSize,
+  },
   width: { property: 'width', transform: transformSize },
   minWidth: { property: 'minWidth', transform: transformSize },
   maxWidth: { property: 'maxWidth', transform: transformSize },