+++ packages/gamut/src/Box/FlexBox.tsx
@@ -1,9 +1,10 @@
+import { styledConfig, system } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
 
-import { Box, BoxProps } from './Box';
+import { boxProps, FlexBoxProps, flexVariants } from './props';
 
-export const FlexBox = styled(Box)<BoxProps>();
-
-FlexBox.defaultProps = {
-  display: 'flex',
-} as BoxProps;
+export const FlexBox = styled('div', styledConfig)<FlexBoxProps>(
+  system.css({ display: 'flex' }),
+  flexVariants,
+  boxProps
+);