+++ packages/gamut/src/NotificationList/NotificationItem.tsx
@@ -96,7 +96,7 @@ export const NotificationItem: React.FC<NotificationItemProps> = ({
   const notificationContent: ReactElement = (
     <FlexBox zIndex={1} position="relative">
       {renderIcon()}
-      <Box flexBasis={0} flexGrow={1} paddingLeft={12}>
+      <Box flex={1} pl={12}>
         <Text id={notificationItemId} fontSize={14} textColor="navy">
           {text}
         </Text>
@@ -121,16 +121,16 @@ export const NotificationItem: React.FC<NotificationItemProps> = ({
   );
 
   const separatorPattern = (
-    <Box paddingX={32} margin={0} aria-hidden="true">
-      <Pattern name="dotsDense" height="1px" display="flex" />
+    <Box px={32} m={0} aria-hidden="true">
+      <Pattern name="checkerDense" height="1px" display="flex" />
     </Box>
   );
 
   return (
     <li>
       <FlexBox
-        paddingY={24}
-        paddingX={32}
+        py={24}
+        px={32}
         alignItems="flex-start"
         justifyContent="space-between"
         position="relative"