+++ packages/styleguide/stories/Foundations/ColorMode/examples.tsx
@@ -1,13 +1,13 @@
 /* eslint-disable local-rules/gamut-import-paths */
-import { Box, FlexBox, Text, Toggle } from '@codecademy/gamut/src';
+import { Anchor, Box, FlexBox, Text, Toggle } from '@codecademy/gamut/src';
 import { Background, ColorMode } from '@codecademy/gamut-styles/src';
 import React, { ComponentProps, useState } from 'react';
 
 export const ColorModeExample = () => {
   const [isDark, setIsDark] = useState(false);
   return (
-    <Box marginY={16} marginBottom={32}>
-      <FlexBox fontWeight="title" marginBottom={8} alignItems="center">
+    <Box mt={16} mb={32}>
+      <FlexBox fontWeight="title" mb={8} alignItems="center">
         <Text marginRight={8} as="strong">
           Use Dark Mode
         </Text>
@@ -20,21 +20,15 @@ export const ColorModeExample = () => {
         />
       </FlexBox>
       <ColorMode mode={isDark ? 'dark' : 'light'}>
-        <Box
-          backgroundColor="background"
-          padding={24}
-          borderColor="text"
-          borderStyle="solid"
-          borderWidth="1px"
-        >
+        <Box bg="background" p={24} border={1}>
           <Text as="h5" fontSize={26} fontWeight="title" marginBottom={16}>
             {isDark ? 'Dark' : 'Light'} Mode
           </Text>
           <Text as="p" marginBottom={16}>
             Lorem ipsum dolor sit amet, sed do eiusmod tempor incididunt ut
-            labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-            exercitation ullamco laboris nisi ut aliquip ex ea commodo
-            consequat.
+            labore et dolore <Anchor>magna aliqua</Anchor>. Ut enim ad minim
+            veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex
+            ea commodo consequat.
           </Text>
         </Box>
       </ColorMode>
@@ -53,8 +47,9 @@ export const BackgroundExample: React.FC<ComponentProps<typeof Background>> = ({
       </Text>
       <Text as="p" marginBottom={16}>
         Lorem ipsum dolor sit amet, sed do eiusmod tempor incididunt ut labore
-        et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud
-        exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
+        et dolore <Anchor>magna aliqua</Anchor>. Ut enim ad minim veniam, quis
+        nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
+        consequat.
       </Text>
       {children}
     </Background>